backend/requirements.txt
fastapi
uvicorn
deap
simpy
psycopg2-binary
pydantic
